Fix state assignment, make TS happy
State objects were nested as a self-named key instead of being directly assigned to root keys.
TS was sad because the initial state was null, and we were assigning an object to that implied constraint.

This can be improved by defining types on the AppContext object and state objects, but this is OK for now.